requirements.txt
asgiref==3.5.2
Django==4.0.4
django-ckeditor==6.4.1
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
djangorestframework==3.12.4
psycopg2==2.8.6
Pillow==8.4.0
pytz==2021.1
gunicorn==20.1.0
whitenoise==6.1.0
dj_database_url==0.5.0
